fn:transform - refactor into own package
The fn:transform implementation is large, and has accumulated many static and dynamic subclasses of FnTransform. This change creates its own package, and creates several top level classes in this package.

The FnTransform module itself becomes a much smaller stub around a call into a Transform object which is part of the new package.
