+++ exist-core/src/main/java/org/exist/xquery/functions/fn/transform/TreeUtils.java
@@ -20,7 +20,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.exist.xquery.functions.fn;
+package org.exist.xquery.functions.fn.transform;
 
 import net.sf.saxon.s9api.XdmNode;
 import org.exist.xquery.value.NodeValue;
